credit-pack/pyproject.toml
[project]
name = "credit-pack"
version = "0.1.0"
description = "Credit Pack multi-agent — ADK Samples style (deal analysis, requirements, compliance, structure, drafting)"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "google-adk>=1.14",
    "python-dotenv>=1.0.1",
    "google-cloud-aiplatform[adk]>=1.93.0",
    "google-genai>=1.0.0",
    "python-docx>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["credit_pack"]
